hnswlib, refactor: change labeltype type to int32_t.

+++ hnswlib/hnswlib.h
@@ -113,7 +113,7 @@ static bool AVX512Capable() {
 #include <string.h>
 
 namespace hnswlib {
-typedef size_t labeltype;
+typedef int32_t labeltype;
 
 // This can be extended to store state for filtering (e.g. from a std::set)
 class BaseFilterFunctor {
